Make registers fall in line
Registers now have the same files/accessor/string functions as sysregs
and operations.
