+++ CMakeLists.txt
@@ -19,6 +19,7 @@ file(GLOB SOURCES
 	disassembler/decode.c
 	disassembler/format.c
 	disassembler/sysregs.c
+	disassembler/regs.c
 	disassembler/encodings_dec.c
 	disassembler/encodings_fmt.c
 	disassembler/operations.c
